# backend/documentation/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/communities/{communityId}/discussions/count": {
      "get": {
        "operationId": "DiscussionsController_countDiscussionsAfterDate",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "description": "Community ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "addedAfterDate",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Number of discussions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountDiscussionDto"
                }
              }
            }
          }
        },
        "summary": "Count discussions added after a specific date",
        "tags": [
          "Discussionss"
        ]
      }
    },
    "/api/communities/{communityId}/discussions/messages/count": {
      "get": {
        "operationId": "DiscussionsController_countMessagesAfterDate",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "description": "Community ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "postedAfterDate",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Number of messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountDiscussionMessageDto"
                }
              }
            }
          }
        },
        "summary": "Count messages posted after a specific date",
        "tags": [
          "Discussionss"
        ]
      }
    },
...
  },
  "info": {
    "title": "@edifice.io/community-backend",
    "description": "API description of community backend",
    "version": "1.0.0-develop-pedago.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CountDiscussionDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number"
          }
        },
        "required": [
          "count"
        ]
      },
      "CountDiscussionMessageDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number"
          }
        },
        "required": [
          "count"
        ]
      },
...
    }
  }
}